Refactor BookmarksAdapter
